Levels 2 and 3 First Draft
Levels 2 and 3 have been implemented as new scenes. Additionally, speed modifiers have been added to the player controller code. A new tag and pickup called Speed Boost has been added (which modify player speed for a duration of 1.5 seconds). Additionally, the ball has a new physic material called Ball in level 2 and 3 to aid in creating the feel of the mud material.